Usage of API
This file uses simple open source Pokebase API. User enters Pokemon ID or Pokemon name and the output shows what is the name, what is the id, what are the abilities and where do you find your target.  In a scenario when user searches for a number or copy pastes non existing link/ pokemon name, the program will exit 
because status.code is different than 200.